Experiment with Isend.
