+++ cpp/dolfinx/graph/GraphBuilder.h
@@ -10,7 +10,6 @@
 #include <array>
 #include <cstdint>
 #include <dolfinx/common/MPI.h>
-#include <dolfinx/common/types.h>
 #include <dolfinx/mesh/cell_types.h>
 #include <tuple>
 #include <utility>
