+++ cpp/dolfinx/mesh/CMakeLists.txt
@@ -9,6 +9,7 @@ set(HEADERS_mesh
   ${CMAKE_CURRENT_SOURCE_DIR}/MeshIterator.h
   ${CMAKE_CURRENT_SOURCE_DIR}/Ordering.h
   ${CMAKE_CURRENT_SOURCE_DIR}/Partitioning.h
+  ${CMAKE_CURRENT_SOURCE_DIR}/PartitioningNew.h
   ${CMAKE_CURRENT_SOURCE_DIR}/MeshQuality.h
   ${CMAKE_CURRENT_SOURCE_DIR}/Topology.h
   ${CMAKE_CURRENT_SOURCE_DIR}/MeshValueCollection.h
@@ -26,6 +27,7 @@ target_sources(dolfinx PRIVATE
   ${CMAKE_CURRENT_SOURCE_DIR}/Geometry.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/Ordering.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/Partitioning.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/PartitioningNew.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/MeshQuality.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/Topology.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/PartitionData.cpp
