+++ cpp/dolfinx/mesh/Partitioning.cpp
@@ -4,13 +4,15 @@
 //
 // SPDX-License-Identifier:    LGPL-3.0-or-later
 
-#include "Partitioning.h"
+#include <mpi.h>
+
 #include "DistributedMeshTools.h"
 #include "Mesh.h"
 #include "MeshEntity.h"
 #include "MeshFunction.h"
 #include "MeshValueCollection.h"
 #include "PartitionData.h"
+#include "Partitioning.h"
 #include "Topology.h"
 #include <algorithm>
 #include <cmath>
@@ -484,456 +486,6 @@ void distribute_cell_layer(
 
 } // namespace
 
-//-----------------------------------------------------------------------------
-std::vector<int>
-Partitioning::partition_cells(MPI_Comm comm, int nparts,
-                              const mesh::CellType cell_type,
-                              const graph::AdjacencyList<std::int64_t>& cells)
-{
-  LOG(INFO) << "Compute partition of cells across processes";
-
-  // FIXME: Update GraphBuilder to use AdjacencyList
-  // Wrap AdjacencyList
-  const Eigen::Map<const Eigen::Array<std::int64_t, Eigen::Dynamic,
-                                      Eigen::Dynamic, Eigen::RowMajor>>
-      _cells(cells.array().data(), cells.num_nodes(),
-             mesh::num_cell_vertices(cell_type));
-
-  // Compute dual graph (for the cells on this process)
-  const auto [local_graph, graph_info]
-      = graph::GraphBuilder::compute_dual_graph(comm, _cells, cell_type);
-  const auto [num_ghost_nodes, num_local_edges, num_nonlocal_edges]
-      = graph_info;
-
-  // Build graph
-  graph::CSRGraph<SCOTCH_Num> csr_graph(comm, local_graph);
-  std::vector<std::size_t> weights;
-
-  // Call partitioner
-  const auto [partition, ignore] = graph::SCOTCH::partition(
-      comm, (SCOTCH_Num)nparts, csr_graph, weights, num_ghost_nodes);
-
-  return partition;
-}
-//-----------------------------------------------------------------------------
-std::tuple<graph::AdjacencyList<std::int32_t>,
-           std::map<std::int64_t, std::int32_t>, std::int32_t>
-Partitioning::create_local_adjacency_list(
-    const graph::AdjacencyList<std::int64_t>& cells)
-{
-  const Eigen::Array<std::int64_t, Eigen::Dynamic, 1>& array = cells.array();
-  std::vector<std::int32_t> array_local(array.rows());
-
-  // Re-map global to local
-  int local = 0;
-  std::map<std::int64_t, std::int32_t> global_to_local;
-  for (int i = 0; i < array.rows(); ++i)
-  {
-    const std::int64_t global = array(i);
-    auto it = global_to_local.find(global);
-    if (it == global_to_local.end())
-    {
-      array_local[i] = local;
-      global_to_local.insert({global, local});
-      ++local;
-    }
-    else
-      array_local[i] = it->second;
-  }
-
-  // FIXME: Update AdjacencyList to avoid this
-  const Eigen::Array<std::int32_t, Eigen::Dynamic, 1>& offsets
-      = cells.offsets();
-  std::vector<std::int32_t> _offsets(offsets.data(),
-                                     offsets.data() + offsets.rows());
-  return {graph::AdjacencyList<std::int32_t>(array_local, _offsets),
-          global_to_local, local};
-}
-//-----------------------------------------------------------------------------
-void Partitioning::create_distributed_adjacency_list(
-    MPI_Comm comm, const mesh::Topology& topology_local,
-    const std::map<std::int64_t, std::int32_t>& global_to_local_vertices)
-{
-  // Create distributed cell-vertex connectivity
-  //   1. New AdjacencyList
-  //   2. IndexMap for vertices
-
-  // Get list of boundary vertices
-  const int dim = topology_local.dim();
-  auto facet_cell = topology_local.connectivity(dim - 1, dim);
-  if (!facet_cell)
-  {
-    throw std::runtime_error(
-        "Need facet-cell connectivity to build distributed adjacency list.");
-  }
-
-  auto facet_vertex = topology_local.connectivity(dim - 1, 0);
-  if (!facet_vertex)
-  {
-    throw std::runtime_error(
-        "Need facet-vertex connectivity to build distributed adjacency list.");
-  }
-
-  auto map_vertex = topology_local.index_map(0);
-  if (!map_vertex)
-    throw std::runtime_error("Need vertex IndexMap from topology.");
-  assert(map_vertex->num_ghosts() == 0);
-  std::vector<bool> exterior_vertex(map_vertex->size_local(), false);
-  for (int f = 0; f < facet_cell->num_nodes(); ++f)
-  {
-    if (facet_cell->num_links(f) == 1)
-    {
-      auto vertices = facet_vertex->links(f);
-      for (int j = 0; j < vertices.rows(); ++j)
-        exterior_vertex[vertices[j]] = true;
-    }
-  }
-
-  // Get maximum global index
-  std::int64_t my_max_global_index = 0;
-  if (!global_to_local_vertices.empty())
-    my_max_global_index = global_to_local_vertices.rbegin()->first;
-  const std::int64_t max_global_index
-      = dolfinx::MPI::all_reduce(comm, my_max_global_index, MPI_MAX);
-
-  // std::cout << "Max my index: " << my_max_global_index << std::endl;
-  // std::cout << "Max index:    " << max_global_index << std::endl;
-
-  const int size = dolfinx::MPI::size(comm);
-  std::vector<int> number_to_send(size, 0);
-  for (auto vertex : global_to_local_vertices)
-  {
-    if (exterior_vertex[vertex.second])
-    {
-      const int owner
-          = dolfinx::MPI::index_owner(comm, vertex.first, max_global_index + 1);
-      number_to_send[owner] += 1;
-    }
-  }
-
-  // Compute global vertex send offsets
-  std::vector<int> disp_send(size + 1, 0);
-  std::partial_sum(number_to_send.begin(), number_to_send.end(),
-                   disp_send.begin() + 1);
-
-  // Pack global vertex send data
-  std::vector<int> disp_tmp = disp_send;
-  std::vector<std::int64_t> vertices_send(disp_tmp.back());
-  for (auto vertex : global_to_local_vertices)
-  {
-    if (exterior_vertex[vertex.second])
-    {
-      const int owner
-          = dolfinx::MPI::index_owner(comm, vertex.first, max_global_index + 1);
-      vertices_send[disp_tmp[owner]++] = vertex.first;
-    }
-  }
-
-  // Send/receive number of vertex indices to communicate to each process
-  std::vector<int> number_to_recv(size);
-  MPI_Alltoall(number_to_send.data(), 1, MPI_INT, number_to_recv.data(), 1,
-               MPI_INT, comm);
-
-  // Compute receive displacements
-  std::vector<int> disp_recv(size + 1, 0);
-  std::partial_sum(number_to_recv.begin(), number_to_recv.end(),
-                   disp_recv.begin() + 1);
-
-  // Send/receive global indices
-  std::vector<std::int64_t> vertices_recv(disp_recv.back());
-  MPI_Alltoallv(vertices_send.data(), number_to_send.data(), disp_send.data(),
-                MPI::mpi_type<std::int64_t>(), vertices_recv.data(),
-                number_to_recv.data(), disp_recv.data(),
-                MPI::mpi_type<std::int64_t>(), comm);
-
-  // Assign ownership to lowest rank
-  // FIXME: revise this to balance load
-  const std::array<std::int64_t, 2> range
-      = dolfinx::MPI::local_range(comm, max_global_index + 1);
-  std::vector<int> marker(range[1] - range[0], -1);
-  std::vector<int> owner_send(disp_recv.back(), -1);
-  for (int i = 0; i < size; ++i)
-  {
-    for (int j = disp_recv[i]; j < disp_recv[i + 1]; ++j)
-    {
-      // Get back to zero reference index
-      const std::int64_t index = vertices_recv[j] - range[0];
-      if (marker[index] < 0)
-      {
-        owner_send[j] = i;
-        marker[index] = i;
-      }
-      else
-        owner_send[j] = marker[index];
-    }
-  }
-
-  // if (dolfinx::MPI::rank(comm) == 1)
-  // {
-  //   for (std::size_t i = 0; i < owner_send.size(); ++i)
-  //     std::cout << "pre: " << i << ", " << owner_send[i] << std::endl;
-
-  //   // std::cout << "Size check (A): " << owner_send.size() << ", " <<
-  //   disp_recv
-  // }
-
-  // Send/receive owner rank (transpose of send/receive global indices)
-  // for each entry in vertices_send
-  std::vector<int> owner_recv(disp_send.back(), -1);
-  MPI_Alltoallv(owner_send.data(), number_to_recv.data(), disp_recv.data(),
-                MPI::mpi_type<int>(), owner_recv.data(), number_to_send.data(),
-                disp_send.data(), MPI::mpi_type<int>(), comm);
-  assert(std::count(owner_recv.begin(), owner_recv.end(), -1) == 0);
-
-  std::map<std::int64_t, std::int32_t> global_to_local_owned0;
-  for (auto vertex : global_to_local_vertices)
-  {
-    if (!exterior_vertex[vertex.second])
-      global_to_local_owned0.insert(vertex);
-  }
-
-  const int rank = dolfinx::MPI::rank(comm);
-  std::map<std::int64_t, std::int32_t> global_to_local_owned1;
-  std::map<std::int64_t, std::int32_t> global_to_local_ghost;
-  std::map<std::int64_t, std::int32_t> tmp_map;
-  assert(owner_recv.size() == vertices_send.size());
-  for (std::size_t i = 0; i < owner_recv.size(); ++i)
-  {
-    auto it = global_to_local_vertices.find(vertices_send[i]);
-    assert(it != global_to_local_vertices.end());
-    if (owner_recv[i] == rank)
-      global_to_local_owned1.insert({vertices_send[i], it->second});
-    else
-      global_to_local_ghost.insert({vertices_send[i], it->second});
-
-    tmp_map.insert({vertices_send[i], it->second});
-  }
-
-  // if (rank == 1)
-  // {
-  //   std::cout << "Test: " << map_vertex->size_local() << ", "
-  //             << facet_vertex->array().maxCoeff() << std::endl;
-  //   for (auto v : global_to_local_vertices)
-  //     std::cout << "   Local: " << v.second << std::endl;
-  // }
-  // return;
-
-  // Number owned vertices locally
-  std::vector<std::int32_t> local_to_local_new(map_vertex->size_local(), -1);
-  std::int32_t p = 0;
-  for (auto it = global_to_local_owned0.begin();
-       it != global_to_local_owned0.end(); ++it)
-  {
-    // if (rank == 0)
-    //   std::cout << "Gl: " << it->first << ", " << p << std::endl;
-    assert(it->second < (int)local_to_local_new.size());
-    local_to_local_new[it->second] = p++;
-    // local_to_local_new[p++] = it->second;
-  }
-  for (auto it = global_to_local_owned1.begin();
-       it != global_to_local_owned1.end(); ++it)
-  {
-    // if (rank == 0)
-    //   std::cout << "Gl: " << it->first << ", " << p << std::endl;
-    assert(it->second < (int)local_to_local_new.size());
-    local_to_local_new[it->second] = p++;
-    // local_to_local_new[p++] = it->second;
-  }
-
-  // Compute process offset
-  const std::int64_t num_owned_vertices
-      = global_to_local_owned0.size() + global_to_local_owned1.size();
-  const std::int64_t offset_global
-      = dolfinx::MPI::global_offset(comm, num_owned_vertices, true);
-  std::cout << "My offset: " << rank << ", " << offset_global << std::endl;
-
-  // Send/receive new global indices for ghosts
-  // - global_to_local_owned1 has global -> local old
-  // - local_to_local_new has local old -> local new
-
-  // - Send old global/new index to owner (owner_recv is owner for each
-  //   vertex in vertices_send)
-  // - Owner sends back new global index
-
-  // Send/receive new global indices
-  std::vector<std::int64_t> vertices_new_send(vertices_send.size(), -1);
-  for (std::size_t i = 0; i < vertices_send.size(); ++i)
-  {
-    auto it = tmp_map.find(vertices_send[i]);
-    if (it != tmp_map.end())
-    {
-      assert(i < vertices_new_send.size());
-      assert(it->second < (int)local_to_local_new.size());
-
-      if (local_to_local_new[it->second] >= 0)
-        vertices_new_send[i] = local_to_local_new[it->second] + offset_global;
-
-      // if (rank == 1)
-      //   std::cout << "Test: " << rank << ", " << it->first << ", "
-      //             << vertices_new_send[i] << std::endl;
-    }
-    if (rank == 1)
-      std::cout << "Test: " << rank << ", " << it->first << ", "
-                << vertices_new_send[i] << std::endl;
-  }
-
-  std::vector<std::int64_t> vertices_new_recv(disp_recv.back());
-  MPI_Alltoallv(vertices_new_send.data(), number_to_send.data(),
-                disp_send.data(), MPI::mpi_type<std::int64_t>(),
-                vertices_new_recv.data(), number_to_recv.data(),
-                disp_recv.data(), MPI::mpi_type<std::int64_t>(), comm);
-
-  std::vector<std::int64_t> vowner_send(disp_recv.back(), -1);
-
-  for (int i = 0; i < size; ++i)
-  {
-    for (int j = disp_recv[i]; j < disp_recv[i + 1]; ++j)
-    {
-      if (rank == 1)
-        std::cout << "Recv (A1): " << i << ", " << vertices_new_recv[j]
-                  << std::endl;
-      if (vertices_new_recv[j] >= 0)
-        vowner_send[j] = vertices_new_recv[j];
-      // if (rank == 0)
-      //   std::cout << "Send (A2): " << i << ", " << vowner_send[j] << std::endl;
-    }
-  }
-
-
-  // const std::array<std::int64_t, 2> range
-  //     = dolfinx::MPI::local_range(comm, max_global_index + 1);
-  // std::vector<int> marker(range[1] - range[0], -1);
-  // std::vector<int> owner_send(disp_recv.back(), -1);
-  // for (int i = 0; i < size; ++i)
-  // {
-  //   for (int j = disp_recv[i]; j < disp_recv[i + 1]; ++j)
-  //   {
-  //     // Get back to zero reference index
-  //     const std::int64_t index = vertices_recv[j] - range[0];
-  //     if (marker[index] < 0)
-  //     {
-  //       owner_send[j] = i;
-  //       marker[index] = i;
-  //     }
-  //     else
-  //       owner_send[j] = marker[index];
-  //   }
-  // }
-
-
-  // if (rank == 0)
-  // {
-  //   for (auto e : vowner_send)
-  //     std::cout << "e: " << e << std::endl;
-  //   // for (auto e : global_to_local_ghost)
-  //   //   std::cout << "g: " << e.first << std::endl;
-  // }
-
-  std::vector<std::int64_t> vertices_final_recv(disp_send.back());
-  MPI_Alltoallv(vowner_send.data(), number_to_recv.data(), disp_recv.data(),
-                MPI::mpi_type<std::int64_t>(), vertices_final_recv.data(),
-                number_to_send.data(), disp_send.data(),
-                MPI::mpi_type<std::int64_t>(), comm);
-
-  if (rank == 1)
-  {
-    std::cout << "------------" << std::endl;
-    for (int i = 0; i < vertices_final_recv.size(); ++i)
-    {
-      std::cout << "Old/new: " << vertices_send[i] << ", "
-                << vertices_final_recv[i] << std::endl;
-    }
-  }
-  // for (int i = 0; i < size; ++i)
-  // {
-  //   for (int j = disp_send[i]; j < disp_send[i + 1]; ++j)
-  //   {
-  //     if (rank == 1)
-  //       std::cout << "Recv (B): " << i << ", " << vertices_final_recv[j]
-  //                 << std::endl;
-  //   }
-  // }
-
-  // if (rank == 1)
-  // {
-  //   for (auto e : vertices_final_recv)
-  //     std::cout << "e: " << e << std::endl;
-  //   for (auto e : global_to_local_ghost)
-  //     std::cout << "g: " << e.first << std::endl;
-  // }
-  // assert(std::count(vowner_send.begin(), vowner_send.end(), -1) == 0);
-}
-//-----------------------------------------------------------------------------
-std::pair<graph::AdjacencyList<std::int64_t>, std::vector<int>>
-Partitioning::distribute(const MPI_Comm& comm,
-                         const graph::AdjacencyList<std::int64_t>& list,
-                         const std::vector<int>& owner)
-{
-  const int size = dolfinx::MPI::size(comm);
-
-  // Compute number of links to send to each process
-  std::vector<int> num_per_dest_send(size, 0);
-  assert(list.num_nodes() == (int)owner.size());
-  for (int i = 0; i < list.num_nodes(); ++i)
-    num_per_dest_send[owner[i]] += list.num_links(i) + 1;
-
-  // Compute send array displacements
-  std::vector<int> disp_send(size + 1, 0);
-  std::inclusive_scan(num_per_dest_send.begin(), num_per_dest_send.end(),
-                      disp_send.begin() + 1);
-
-  // Send/receive number of items to communicate
-  std::vector<int> num_per_dest_recv(size, 0);
-  MPI_Alltoall(num_per_dest_send.data(), 1, MPI_INT, num_per_dest_recv.data(),
-               1, MPI_INT, comm);
-
-  // Compite receive array displacements
-  std::vector<int> disp_recv(size + 1, 0);
-  std::inclusive_scan(num_per_dest_recv.begin(), num_per_dest_recv.end(),
-                      disp_recv.begin() + 1);
-
-  // Prepare send buffer
-  std::vector<int> offset = disp_send;
-  std::vector<std::int64_t> data_send(list.array().rows() + list.num_nodes());
-  for (int i = 0; i < list.num_nodes(); ++i)
-  {
-    const int dest = owner[i];
-    auto links = list.links(i);
-    data_send[offset[dest]] = links.rows();
-    ++offset[dest];
-    for (int j = 0; j < links.rows(); ++j)
-      data_send[offset[dest] + j] = links(j);
-    offset[dest] += links.rows();
-  }
-
-  // Send/receive data
-  std::vector<std::int64_t> data_recv(disp_recv.back());
-  MPI_Alltoallv(data_send.data(), num_per_dest_send.data(), disp_send.data(),
-                MPI_INT64_T, data_recv.data(), num_per_dest_recv.data(),
-                disp_recv.data(), MPI_INT64_T, comm);
-
-  // Unpack receive buffer
-  std::vector<std::int64_t> array;
-  std::vector<std::int32_t> list_offset(1, 0);
-  std::vector<int> src;
-  for (std::size_t p = 0; p < disp_recv.size() - 1; ++p)
-  {
-    for (int i = disp_recv[p]; i < disp_recv[p + 1]; ++i)
-    {
-      src.push_back(p);
-      const std::int64_t num_links = data_recv[i];
-      const int pos = i;
-      for (int j = 1; j <= num_links; ++j)
-      {
-        array.push_back(data_recv[pos + j]);
-        i += 1;
-      }
-      list_offset.push_back(list_offset.back() + num_links);
-    }
-  }
-
-  return {graph::AdjacencyList<std::int64_t>(array, list_offset), src};
-}
 //-----------------------------------------------------------------------------
 std::tuple<
     std::shared_ptr<common::IndexMap>, std::vector<std::int64_t>,
